backend/requirements.txt
# FastAPI & Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.4

# AI & ML
langchain==0.1.4
langchain-openai==0.0.5
langchain-google-genai==0.0.6
langgraph==0.0.20
openai==1.10.0
tiktoken==0.5.2

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
openpyxl==3.1.2
pillow==10.2.0
pytesseract==0.3.10

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
aiofiles==23.2.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
